requirements.txt
requests
bs4
feedly-client>=0.22